Added branches and commit info to git status results
This change updates the status function to include information about the
tracking branch, the current branch, number of commits ahead, and number
of commits behind. Now, the response for status will look as follows:

{ ahead: '1',
  behind: '3',
  current: 'current_branch',
  tracking: 'origin/master',
  not_added: [],
  deleted: [],
  modified: [],
  created: [],
  conflicted: [] }

This information is attained by adding '-b' to the status command, which
prints out branch and commit information in a parsable format.