Feature/timeout plugin (#591)
feat: Timeout Plugin

The Timeout Plugin can be used to shut down the spawned `git` processes if they stop responding for a supplied time (in `ms`). The timeout is reset each time `git` outputs to either `stdOut` or `stdErr. When the timeout is reached, the `git` process is killed with message `SIGINT` - the interrupt signal which can be gracefully handled if necessary.

When a task is killed in this way, the `simple-git` task that called it will fail (rejecting the promise it returned and calling back with an `Error` if the trailing callback was supplied). The error message will include `TimeoutPlugin` and detail of the timeout that expired.

feat: convert `git-factory` to TypeScript

- Force closing a `git` process now emits a `GitPluginError` with detail of the timeout
- Move plugin documentation to separate markdowns in docs folder