feat: task callback types defined as single function type
The TypeScript `Parameters` type doesn't permit creating a type that is the discriminating union of the argument tuples, meaning the `SimpleGitTaskCallback` was always typed as the error state `(err: E) => void`.

To resolve this, the `SimpleGitTaskCallback` is now typed with a single function type in the same way as the official `node` types: `(err: null | E, data: T) => void`.